backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.0.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
